requirements.txt
Authlib==1.3.2
blinker==1.8.2
cachelib==0.13.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
ecdsa==0.19.0
Flask==3.0.3
Flask-Session==0.8.0
Flask-WTF==1.2.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msgspec==0.18.6
msal==1.31.0
packaging==24.1
pyasn1==0.6.1
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
python-jose==3.3.0
requests==2.32.3
rsa==4.9
six==1.16.0
urllib3==2.2.3
Werkzeug==3.0.4
WTForms==3.1.2
